requirements.txt
requests>=2.32.3
selenium==4.15.2
crewai==0.119.0
crewai-tools==0.45.0
robotexclusionrulesparser==1.7.1
webdriver-manager==4.0.1
pdfplumber==0.11.4
sentence-transformers==2.2.2
google-generativeai==0.8.3
transformers==4.36.0
python-dotenv==1.0.0
tqdm==4.66.1
langchain==0.3.25
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
langchain-community>=0.0.10
gradio==5.1.0
pinecone-client==3.0.3
markdown2==2.4.10